requirements.txt
discord.py[voice]
substring
requests
nltk
websockets